feat: move table files in shared dirs
